3-Weather-Inf- Agent/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   mcp
    #   openai
    #   sse-starlette
    #   starlette
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
certifi==2025.7.14
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   griffe
    #   tqdm
distro==1.9.0
    # via openai
griffe==1.8.0
    # via openai-agents
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   mcp
    #   openai
httpx-sse==0.4.1
    # via mcp
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.10.0
    # via openai
jsonschema==4.25.0
    # via mcp
jsonschema-specifications==2025.4.1
    # via jsonschema
mcp==1.12.2
    # via openai-agents
openai==1.97.1
    # via openai-agents
openai-agents==0.2.3
    # via 3-weather-inf-agent (pyproject.toml)
pydantic==2.11.7
    # via
    #   mcp
    #   openai
    #   openai-agents
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via mcp
python-decouple==3.8
    # via 3-weather-inf-agent (pyproject.toml)
python-dotenv==1.1.1
    # via pydantic-settings
python-multipart==0.0.20
    # via mcp
pywin32==311
    # via mcp
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.4
    # via openai-agents
rpds-py==0.26.0
    # via
    #   jsonschema
    #   referencing
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sse-starlette==2.4.1
    # via mcp
starlette==0.47.2
    # via mcp
tqdm==4.67.1
    # via openai
types-requests==2.32.4.20250611
    # via openai-agents
typing-extensions==4.14.1
    # via
    #   openai
    #   openai-agents
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via
    #   requests
    #   types-requests
uvicorn==0.35.0
    # via mcp
